Add ability to save and improve printing
Add ability to save todo list to a file. This is the first iteration and
as such the file name is hardcoded. Saving the file allowed for cleaning
up the printing of todo lists a little bit since the code can be shared
amongs the two.

Future work:
* Extract printing logic into its own file
* Extract saving logic into its own file

We need to